FIX another ambiguous error on the draw patttern

+++ projects/07-led_matrix_and_shift_register/code/libs/LedMatrixAndShiftRegisterDrawer/LedMatrixAndShiftRegisterDrawer.cpp
@@ -77,6 +77,6 @@ void LedMatrixAndShiftRegisterDrawer::drawAnimation(const uint8_t* const animati
 {
     for (uint8_t pattern = 0; pattern < patternCount; ++pattern)
     {
-        drawPattern(animation[pattern], frameTimeMs); // render one frame
+        drawPattern(animation[pattern], (unsigned long)frameTimeMs); // render one frame
     }
 }
